filtering!: a page-size clamp a decoder can reach, and a ceiling a consumer can set (#299)
* filtering: the page-size clamp a wire decoder can reach, so it clamps before it narrows

The ceiling has to be applied to the wide value, before the narrowing to
uint16, and every wire format narrows first because it has no choice:
protobuf has no uint16 so a page size crosses as a uint32, JSON hands a
decoder a number, a query parameter hands it a string. All three reach
*uint16 before there is a QueryFilter to hold one, which is before
Normalize can see the value — and a narrowing that happens first is
silent. 70000 wraps to 4464, Normalize clamps that to 250, and the client
receives a legible-looking answer to a question it did not ask.

FromParams got this right by being inside the package and calling the
unexported clamp on the uint64 it parsed. Every decoder outside the
package was on its own, and restated the ceiling by hand where this
package cannot see it drift.

Exports both halves. ClampResponseSize is the existing function unchanged,
for a parser that has the wide value before it has a filter to put it in.
SetMaxResponseSize is that applied to a filter, mirroring SetCursor, and
is what a decoder holding one should reach for: an exported clamp is still
something a decoder can forget to call, whereas a setter taking uint64 has
no wrong order left to get right.

Closes #295.


* filtering: a page-size ceiling a consumer can set, and a schema that follows it

MaxQueryFilterLimit was a constant, so 250 was every consumer's ceiling
whether or not it was the right one for their rows. Lowering it was
always possible by hand — nothing raises a page size — but raising it
was not possible at all, which left a service that pages cheaply held to
a number this module had no business picking for it.

Makes it a var. The half that is easy to leave behind is the published
bound: `maximum` was a struct tag, and a tag is fixed when this package
compiles, so a service that raised the ceiling would have gone on
promising 250 to every generated client and every tool-calling model
while clamping at something else. That is the drift this package
reflects its schema to avoid, so the tag is gone and PrepareJSONSchema
writes the current value instead — the hook any swaggest reflector
calls, which means QueryFilterSchema here and equally the openapi-go
reflector routing runs over a consumer's own response types, where this
package has no call of its own to patch afterwards.

The reflected document is no longer cached, for the same reason: a
sync.OnceValue froze whichever ceiling was in place the first time
anybody asked. Reflection now costs ~136µs against ~13µs, paid by
callers building a tool definition or a spec at startup.

The var is typed uint16 rather than left untyped, so a ceiling too large
to store in a page size is a compile error at the assignment rather than
a truncation at the clamp. ClampResponseSize still takes a uint64 and
still has to.

DefaultQueryFilterLimit stays a constant, and the doc says why: it is
applied at runtime here, but database/querygen also writes it into SQL
at generate time, so a value changed in a running process would disagree
with statements that shipped before it started.


* dataprivacy: the walk's page size, without a conversion that no longer converts

CollectAll asked for `new(uint16(filtering.MaxQueryFilterLimit))`, which was
necessary while the ceiling was an untyped constant and is redundant now that
this branch makes it a typed `uint16` var — `unconvert` says so, and it is
right.

The removal is mechanical; the conversion read the var at runtime either way.
Worth noting what is not mechanical, since it arrives from the other side and
neither change describes it on its own: the walk now asks for whatever ceiling
the consumer set, rather than the 250 that used to be fixed when the package
compiled. That is what a walk wanting the largest page allowed should do, and
`CollectAll` already judges shortness against the size the store applied rather
than the size it asked for, so a store clamping below the raised ceiling still
ends the walk correctly.


---------

+++ dataprivacy/collect.go
@@ -78,7 +78,7 @@ func CollectAll[T any](
 	}
 
 	filter := filtering.DefaultQueryFilter()
-	filter.MaxResponseSize = new(uint16(filtering.MaxQueryFilterLimit))
+	filter.MaxResponseSize = new(filtering.MaxQueryFilterLimit)
 
 	var out []T
